+++ src/driver/image.rs
@@ -352,12 +352,6 @@ pub struct ImageInfo {
     #[builder(setter(strip_option))]
     pub height: u32,
 
-    /// Specifies the tiling arrangement of the texel blocks in memory.
-    ///
-    /// The default value of `false` indicates a `VK_IMAGE_TILING_OPTIMAL` image.
-    #[builder(default, setter(strip_option))]
-    pub linear_tiling: bool,
-
     /// The number of levels of detail available for minified sampling of the image.
     #[builder(default = "1", setter(strip_option))]
     pub mip_level_count: u32,
@@ -367,6 +361,12 @@ pub struct ImageInfo {
     /// [samples per texel]: https://registry.khronos.org/vulkan/specs/1.3-extensions/html/vkspec.html#primsrast-multisampling
     #[builder(default = "SampleCount::X1", setter(strip_option))]
     pub sample_count: SampleCount,
+
+    /// Specifies the tiling arrangement of the texel blocks in memory.
+    ///
+    /// The default value is [`vk::ImageTiling::OPTIMAL`].
+    #[builder(default = "vk::ImageTiling::OPTIMAL", setter(strip_option))]
+    pub tiling: vk::ImageTiling,
 
     /// The basic dimensionality of the image.
     ///
@@ -402,7 +402,7 @@ impl ImageInfo {
             depth: Some(depth),
             usage: Some(usage),
             flags: None,
-            linear_tiling: None,
+            tiling: None,
             mip_level_count: None,
             array_elements: Some(array_elements),
             sample_count: None,
@@ -543,11 +543,7 @@ impl ImageInfo {
             .mip_levels(self.mip_level_count)
             .array_layers(array_layers)
             .samples(self.sample_count.into_vk())
-            .tiling(if self.linear_tiling {
-                vk::ImageTiling::LINEAR
-            } else {
-                vk::ImageTiling::OPTIMAL
-            })
+            .tiling(self.tiling)
             .usage(self.usage)
             .sharing_mode(vk::SharingMode::CONCURRENT)
             .initial_layout(vk::ImageLayout::UNDEFINED)
@@ -582,6 +578,8 @@ impl From<UninitializedFieldError> for ImageInfoBuilderError {
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct ImageSubresource {
     /// The number of layers for which this subset applies.
+    ///
+    /// The default value of `None` equates to `vk::REMAINING_ARRAY_LAYERS`.
     pub array_layer_count: Option<u32>,
 
     /// The portion of the image for which this subset applies.
@@ -594,6 +592,8 @@ pub struct ImageSubresource {
     pub base_mip_level: u32,
 
     /// The number of mip levels for which this subset applies.
+    ///
+    /// The default value of `None` equates to `vk::REMAINING_MIP_LEVELS`.
     pub mip_level_count: Option<u32>,
 }
 
@@ -615,22 +615,13 @@ impl From<ImageViewInfo> for ImageSubresource {
             aspect_mask: info.aspect_mask,
             base_mip_level: info.base_mip_level,
             base_array_layer: info.base_array_layer,
-            array_layer_count: Some(match info.ty {
-                ImageType::Cube
-                | ImageType::Texture1D
-                | ImageType::Texture2D
-                | ImageType::Texture3D => {
-                    info.array_layer_count.unwrap_or(vk::REMAINING_ARRAY_LAYERS)
-                }
-                ImageType::CubeArray | ImageType::TextureArray1D | ImageType::TextureArray2D => {
-                    info.array_layer_count.unwrap_or(vk::REMAINING_ARRAY_LAYERS)
-                }
-            }),
-            mip_level_count: info.mip_level_count,
+            array_layer_count: Some(info.array_layer_count.unwrap_or(vk::REMAINING_ARRAY_LAYERS)),
+            mip_level_count: Some(info.mip_level_count.unwrap_or(vk::REMAINING_MIP_LEVELS)),
         }
     }
 }
 
+// TODO: Remove this and use vk::ImageType instead
 /// Describes the number of dimensions and array elements of an image.
 #[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
 pub enum ImageType {
@@ -660,6 +651,19 @@ impl ImageType {
             Self::Texture3D => vk::ImageViewType::TYPE_3D,
             Self::TextureArray1D => vk::ImageViewType::TYPE_1D_ARRAY,
             Self::TextureArray2D => vk::ImageViewType::TYPE_2D_ARRAY,
+        }
+    }
+}
+
+impl From<ImageType> for vk::ImageType {
+    fn from(value: ImageType) -> Self {
+        match value {
+            ImageType::Texture1D | ImageType::TextureArray1D => vk::ImageType::TYPE_1D,
+            ImageType::Texture2D
+            | ImageType::TextureArray2D
+            | ImageType::Cube
+            | ImageType::CubeArray => vk::ImageType::TYPE_2D,
+            ImageType::Texture3D => vk::ImageType::TYPE_3D,
         }
     }
 }
@@ -731,23 +735,32 @@ impl Drop for ImageView {
     derive(Clone, Copy, Debug),
     pattern = "owned"
 )]
+#[non_exhaustive]
 pub struct ImageViewInfo {
     /// The number of layers that will be contained in the view.
+    ///
+    /// The default value of `None` equates to `vk::REMAINING_ARRAY_LAYERS`.
+    #[builder(default)]
     pub array_layer_count: Option<u32>,
 
     /// The portion of the image that will be contained in the view.
     pub aspect_mask: vk::ImageAspectFlags,
 
     /// The first array layer that will be contained in the view.
+    #[builder(default)]
     pub base_array_layer: u32,
 
     /// The first mip level that will be contained in the view.
+    #[builder(default)]
     pub base_mip_level: u32,
 
     /// The format and type of the texel blocks that will be contained in the view.
     pub fmt: vk::Format,
 
     /// The number of mip levels that will be contained in the view.
+    ///
+    /// The default value of `None` equates to `vk::REMAINING_MIP_LEVELS`.
+    #[builder(default)]
     pub mip_level_count: Option<u32>,
 
     /// The basic dimensionality of the view.
@@ -758,7 +771,7 @@ impl ImageViewInfo {
     /// Specifies a default view with the given `fmt` and `ty` values.
     #[allow(clippy::new_ret_no_self)]
     pub fn new(format: vk::Format, ty: ImageType) -> ImageViewInfoBuilder {
-        ImageViewInfoBuilder::new(format, ty)
+        ImageViewInfoBuilder::new(format, ty).aspect_mask(format_aspect_mask(format))
     }
 
     /// Takes this instance and returns it with a newly specified `ImageType`.